Move blotter patches to own file
